+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates npm packages that only affect the repo, not the components, for general security maintenance.
+
 ## 4.4.0 (July 8, 2026)
 
 ### Adds